gemini_code_reviewer/requirements.txt
google-generativeai
python-dotenv
colorama